Compositor:Wayland/Nexus to support reslution param
